requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
supabase>=2.4.0
langchain
langgraph
langchain-mcp-adapters
langchain-groq
langchain-anthropic
